Update documentation; Fix MISC flag on DeviceFactory and FeatureFactory

+++ Devices/DeviceFactory.cs
@@ -20,7 +20,7 @@ namespace HomeSeer.PluginSdk.Devices {
                              Relationship = ERelationship.Device,
                              Interface = pluginId
                          };
-            device.Changes.Add(EProperty.Misc, EMiscFlag.ShowValues);
+            device.Changes.Add(EProperty.Misc, (uint) EMiscFlag.ShowValues);
             device.Changes.Add(EProperty.UserAccess, "Any");
             df._device = device;
